+++ RELEASE_NOTES.md
@@ -52,6 +52,8 @@
 
     Note that the parameter `sampling_period` has been renamed to `input_sampling_period` to better distinguish it from the sampling period parameter in the `resampler_config`.
 
+  * Use the `UNIX_EPOCH` as the default alignment.
+
 * `Resampler`
 
   * The `ResamplerConfig` now takes the resampling period as a `timedelta`. The configuration was renamed from `resampling_period_s` to `resampling_period` accordingly.